[#21] feat: BallCount를 별도 컴포넌트로 분리
기존 Stadium 컴포넌트에 종속되어있던 BallCount영역을 별도 컴포넌트로 분리
