+++ FE/src/styles/Stadium.js
@@ -4,34 +4,6 @@ const Container = styled.div`
 	position: relative;
 `;
 
-const BallCountWrapper = styled.div`
-	position: absolute;
-	top: 10px;
-	left: 15px;
-	color: #37f734;
-	font-size: 20px;
-`;
-
-const BallCount = styled.div`
-	display: flex;
-`;
-
-const LightContainer = styled.div`
-	display: flex;
-	align-items: center;
-	margin-left: 5px;
-	margin-bottom: 2px;
-`;
-
-const Light = styled.div`
-	width: 10px;
-	height: 15px;
-	background: #37f734;
-	:not(:last-child) {
-		margin-right: 5px;
-	}
-`;
-
 const InningInfo = styled.div`
 	position: absolute;
 	top: 10px;
@@ -67,13 +39,4 @@ const RecordBtn = styled.button`
 	border: 3px solid #37f734;
 `;
 
-export {
-	Container,
-	BallCountWrapper,
-	BallCount,
-	LightContainer,
-	Light,
-	InningInfo,
-	PitchBtn,
-	RecordBtn,
-};
+export { Container, InningInfo, PitchBtn, RecordBtn };
